Don't throw INVALID_ACCESS_ERR
Per discussion:

http://lists.whatwg.org/htdig.cgi/whatwg-whatwg.org/2011-September/033235.html

This change partially reverts the commit "Rewrite a bunch of general
behavior" of June 29, 3cb594b3 in git and 1464ec42 in hg.

Since the change to expected test results is huge, I obviously didn't
review the whole thing exactly, although git diff
--color-words='([a-z]+|.)' is useful to analyze it more easily.  I
spot-checked it and it seems sane.