Improve test generation process
The new procedure is much simpler:

1) Run gentest.html in Gecko and WebKit.

2) When it finishes, click the submit button in the webpage itself,
providing an appropriate password.

3) Run conformancetest/fixdata to generate data.js.

4) Use git diff to check for changes in expectations.

This also fixes a couple of encoding-related errors in data.js.  I've
checked them with a hex editor and the new versions are correct.

Also, this builds the 1M-ish test output string as a regular JS string
instead of in the DOM, at the advice of people like roc.  This yields an
approximately order-of-magnitude improvement in test generation time,
which is now down to 2-4 minutes on my desktop from like 20.  Seriously!