Don't special-case blockquote class for WebKit
Previously, WebKit would generate blockquotes with
class="webkit-indent-blockquote", so we had to treat those as simple
indentation elements for compatibility.  Annoyingly, this meant also
treating all indentation elements with classes as simple, so we'd just
remove them along with their class.  Now having a class will make an
indentation element non-simple, so we'll refuse to remove it.
<blockquote class=foo> will become <div class=foo> if outdented.  WebKit
no longer generates these classes, so this seems safe enough for the
future.

Also, I don't think I ever special-cased <span class=Apple-style-span>,
which was far more prevalent before these classes were removed.  Props
to rniwa for removing them.

This was more complicated to fix than I expected.  Of course, it creates
nastier markup when these extra classes are present in the input, but
it's better to avoid the special case in the long term.  The idea is we
shouldn't have such markup in the input.  (The visual results are
unchanged in any event.)

Fixes: http://www.w3.org/Bugs/Public/show_bug.cgi?id=13957
Reported-By: Ryosuke Niwa