Stop using make, just use a shell script
There's really no need for make's features when you have one target
that's made from one file.  It occurs to me that I could roll mkspec
into preprocess and do it in Python, which would be nicer, but it really
doesn't matter much.

This also makes sure that the specification data is up-to-date before
making anything.  I could have added this to the makefile, but it really
seemed pointless to not just switch to a shell script.