Cache results for normalizeColor/getLegacyFontSize
This gives a decent speedup to gentest.html in Chrome.  I didn't measure
exactly, but these two functions were showing up high on profiles and
now they've vanished.  The major remaining culprits are things like
movePreservingRanges and setSelectionValue for self, and for total,
things like deleteSelection, isVisible, and restoreStatesAndValues.  Not
so trivial to fix.